# Improve JPEG Block8x8F Intrinsics for Vector128 paths.