… src/ImageSharp/Common/Helpers/Vector128Utilities.cs
… src/ImageSharp/Common/Helpers/Vector512Utilities.cs
… src/ImageSharp/Formats/Jpeg/Components/Block8x8.cs
… src/ImageSharp/Formats/Jpeg/Components/Block8x8F.Generated.cs
… src/ImageSharp/Formats/Jpeg/Components/Block8x8F.Generated.tt
… src/ImageSharp/Formats/Jpeg/Components/Block8x8F.Intrinsic.cs
… src/ImageSharp/Formats/Jpeg/Components/Block8x8F.Round.cs
… src/ImageSharp/Formats/Jpeg/Components/Block8x8F.Vector128.cs
… src/ImageSharp/Formats/Jpeg/Components/Block8x8F.Vector256.cs
… src/ImageSharp/Formats/Jpeg/Components/Block8x8F.cs
… src/ImageSharp/Formats/Jpeg/Components/ColorConverters/JpegColorConverter.GrayScaleVector128.cs
… src/ImageSharp/Formats/Jpeg/Components/ColorConverters/JpegColorConverter.GrayScaleVector256.cs
… src/ImageSharp/Formats/Jpeg/Components/ColorConverters/JpegColorConverter.GrayScaleVector512.cs
… src/ImageSharp/Formats/Jpeg/Components/ColorConverters/JpegColorConverter.YCbCrVector128.cs
… src/ImageSharp/Formats/Jpeg/Components/ColorConverters/JpegColorConverter.YCbCrVector256.cs
… src/ImageSharp/Formats/Jpeg/Components/ColorConverters/JpegColorConverter.YCbCrVector512.cs
… src/ImageSharp/Formats/Jpeg/Components/ColorConverters/JpegColorConverter.YccKVector128.cs
… src/ImageSharp/Formats/Jpeg/Components/ColorConverters/JpegColorConverter.YccKVector256.cs
… src/ImageSharp/Formats/Jpeg/Components/ColorConverters/JpegColorConverter.YccKVector512.cs
… src/ImageSharp/Formats/Jpeg/Components/FloatingPointDCT.Intrinsic.cs
… src/ImageSharp/Formats/Jpeg/Components/FloatingPointDCT.Vector256.cs
… src/ImageSharp/Formats/Jpeg/Components/FloatingPointDCT.cs
… src/ImageSharp/Formats/Jpeg/Components/ScaledFloatingPointDCT.cs
… src/ImageSharp/Formats/Jpeg/Components/ZigZag.Intrinsic.cs
… src/ImageSharp/Formats/Webp/AlphaDecoder.cs
… src/ImageSharp/ImageSharp.csproj
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/BlockOperations/Block8x8F_LoadFromInt16.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/BlockOperations/Block8x8F_Transpose.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/DecodeJpeg.cs
… tests/ImageSharp.Benchmarks/Config.HwIntrinsics.cs