### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
<!-- A description of the changes proposed in the pull-request -->
This PR adds `Vector128` intrinsic implementations to several methods in `Block8x8F` and reimplements `ZigZag` to migrate intrinsics from `Sse` to general `Vector<128>` methods which should provide a good speedup on mobile.

Performance improvements are measurable. 

### Benchmarks

```
BenchmarkDotNet v0.14.0, Windows 11 (10.0.26100.3915)
11th Gen Intel Core i7-11370H 3.30GHz, 1 CPU, 8 logical and 4 physical cores
.NET SDK 9.0.300-preview.0.25177.5
  [Host]             : .NET 8.0.15 (8.0.1525.16413), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  1. No HwIntrinsics : .NET 8.0.15 (8.0.1525.16413), X64 RyuJIT
  2. SSE             : .NET 8.0.15 (8.0.1525.16413), X64 RyuJIT SSE4.2
  3. AVX             : .NET 8.0.15 (8.0.1525.16413), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI

Runtime=.NET 8.0
```
#### Main

```
| Method                              | Job                | EnvironmentVariables                            | Mean       | Error     | StdDev    | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|------------------------------------ |------------------- |------------------------------------------------ |-----------:|----------:|----------:|------:|--------:|-------:|----------:|------------:|
| 'Baseline 4:4:4 Interleaved'        | 1. No HwIntrinsics | DOTNET_EnableHWIntrinsic=0,DOTNET_FeatureSIMD=0 | 107.930 ms | 1.5990 ms | 1.4957 ms |  1.00 |    0.02 |      - |  47.46 KB |        1.00 |
| 'Baseline 4:4:4 Interleaved'        | 2. SSE             | DOTNET_EnableAVX=0                              |  24.525 ms | 0.1969 ms | 0.1842 ms |  0.23 |    0.00 |      - |   47.1 KB |        0.99 |
| 'Baseline 4:4:4 Interleaved'        | 3. AVX             | Empty                                           |   8.838 ms | 0.0784 ms | 0.0733 ms |  0.08 |    0.00 |      - |  47.06 KB |        0.99 |
|                                     |                    |                                                 |            |           |           |       |         |        |           |             |
| 'Baseline 4:2:0 Interleaved'        | 1. No HwIntrinsics | DOTNET_EnableHWIntrinsic=0,DOTNET_FeatureSIMD=0 |  40.974 ms | 0.1971 ms | 0.1844 ms |  1.00 |    0.01 |      - |  15.22 KB |        1.00 |
| 'Baseline 4:2:0 Interleaved'        | 2. SSE             | DOTNET_EnableAVX=0                              |  11.841 ms | 0.0494 ms | 0.0438 ms |  0.29 |    0.00 |      - |  15.14 KB |        0.99 |
| 'Baseline 4:2:0 Interleaved'        | 3. AVX             | Empty                                           |   7.467 ms | 0.0922 ms | 0.0863 ms |  0.18 |    0.00 |      - |  15.13 KB |        0.99 |
|                                     |                    |                                                 |            |           |           |       |         |        |           |             |
| 'Baseline 4:0:0 (grayscale)'        | 1. No HwIntrinsics | DOTNET_EnableHWIntrinsic=0,DOTNET_FeatureSIMD=0 |   8.920 ms | 0.0859 ms | 0.0718 ms |  1.00 |    0.01 |      - |  12.73 KB |        1.00 |
| 'Baseline 4:0:0 (grayscale)'        | 2. SSE             | DOTNET_EnableAVX=0                              |   2.713 ms | 0.0152 ms | 0.0142 ms |  0.30 |    0.00 |      - |  12.72 KB |        1.00 |
| 'Baseline 4:0:0 (grayscale)'        | 3. AVX             | Empty                                           |   1.204 ms | 0.0078 ms | 0.0065 ms |  0.13 |    0.00 | 1.9531 |  12.71 KB |        1.00 |
|                                     |                    |                                                 |            |           |           |       |         |        |           |             |
| 'Progressive 4:2:0 Non-Interleaved' | 1. No HwIntrinsics | DOTNET_EnableHWIntrinsic=0,DOTNET_FeatureSIMD=0 |  74.589 ms | 0.4589 ms | 0.4068 ms |  1.00 |    0.01 |      - |  39.54 KB |        1.00 |
| 'Progressive 4:2:0 Non-Interleaved' | 2. SSE             | DOTNET_EnableAVX=0                              |  20.615 ms | 0.1037 ms | 0.0919 ms |  0.28 |    0.00 |      - |  39.38 KB |        1.00 |
| 'Progressive 4:2:0 Non-Interleaved' | 3. AVX             | Empty                                           |  11.544 ms | 0.0490 ms | 0.0458 ms |  0.15 |    0.00 |      - |  39.35 KB |        1.00 |
```

#### This PR
```
| Method                              | Job                | EnvironmentVariables                            | Mean       | Error     | StdDev    | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|------------------------------------ |------------------- |------------------------------------------------ |-----------:|----------:|----------:|------:|--------:|-------:|----------:|------------:|
| 'Baseline 4:4:4 Interleaved'        | 1. No HwIntrinsics | DOTNET_EnableHWIntrinsic=0,DOTNET_FeatureSIMD=0 | 108.958 ms | 1.3012 ms | 1.2172 ms |  1.00 |    0.02 |      - |  47.46 KB |        1.00 |
| 'Baseline 4:4:4 Interleaved'        | 2. SSE             | DOTNET_EnableAVX=0                              |  13.185 ms | 0.1547 ms | 0.1447 ms |  0.12 |    0.00 |      - |  47.06 KB |        0.99 |
| 'Baseline 4:4:4 Interleaved'        | 3. AVX             | Empty                                           |   8.754 ms | 0.0501 ms | 0.0468 ms |  0.08 |    0.00 |      - |  47.06 KB |        0.99 |
|                                     |                    |                                                 |            |           |           |       |         |        |           |             |
| 'Baseline 4:2:0 Interleaved'        | 1. No HwIntrinsics | DOTNET_EnableHWIntrinsic=0,DOTNET_FeatureSIMD=0 |  41.072 ms | 0.2252 ms | 0.1996 ms |  1.00 |    0.01 |      - |  15.22 KB |        1.00 |
| 'Baseline 4:2:0 Interleaved'        | 2. SSE             | DOTNET_EnableAVX=0                              |   8.928 ms | 0.0815 ms | 0.0722 ms |  0.22 |    0.00 |      - |  15.14 KB |        0.99 |
| 'Baseline 4:2:0 Interleaved'        | 3. AVX             | Empty                                           |   7.399 ms | 0.0449 ms | 0.0398 ms |  0.18 |    0.00 |      - |  15.13 KB |        0.99 |
|                                     |                    |                                                 |            |           |           |       |         |        |           |             |
| 'Baseline 4:0:0 (grayscale)'        | 1. No HwIntrinsics | DOTNET_EnableHWIntrinsic=0,DOTNET_FeatureSIMD=0 |   8.967 ms | 0.0404 ms | 0.0358 ms |  1.00 |    0.01 |      - |  12.73 KB |        1.00 |
| 'Baseline 4:0:0 (grayscale)'        | 2. SSE             | DOTNET_EnableAVX=0                              |   1.723 ms | 0.0079 ms | 0.0070 ms |  0.19 |    0.00 | 1.9531 |  12.71 KB |        1.00 |
| 'Baseline 4:0:0 (grayscale)'        | 3. AVX             | Empty                                           |   1.215 ms | 0.0051 ms | 0.0048 ms |  0.14 |    0.00 | 1.9531 |  12.73 KB |        1.00 |
|                                     |                    |                                                 |            |           |           |       |         |        |           |             |
| 'Progressive 4:2:0 Non-Interleaved' | 1. No HwIntrinsics | DOTNET_EnableHWIntrinsic=0,DOTNET_FeatureSIMD=0 |  74.980 ms | 0.3103 ms | 0.2751 ms |  1.00 |    0.01 |      - |  39.58 KB |        1.00 |
| 'Progressive 4:2:0 Non-Interleaved' | 2. SSE             | DOTNET_EnableAVX=0                              |  14.541 ms | 0.1421 ms | 0.1329 ms |  0.19 |    0.00 |      - |  39.35 KB |        0.99 |
| 'Progressive 4:2:0 Non-Interleaved' | 3. AVX             | Empty                                           |  12.342 ms | 0.2376 ms | 0.2440 ms |  0.16 |    0.00 |      - |  39.35 KB |        0.99 |
```

CC 
@tannergooding - I think I got everything right performance-wise though I have commented with `TODO` where there may be more low hanging fruit.
@beeradmoore - I'm hoping this makes a real difference with the MAUI benchmarks. There were several places where we were falling back to scalar implementations for ARM and WASM.

<!-- Thanks for contributing to ImageSharp! -->