@beeradmoore 2025-05-07
@JimBobSquarePants , when I build the nuget (`dotnet pack -c Release`) to test with it looks like it compiles with .NET 8 SDK. 

<img width="810" alt="Screenshot 2025-05-07 at 5 57 53 pm" src="https://github.com/user-attachments/assets/22db95cf-4f7b-443c-b957-07e55448c057" />

I assume that's what the above says, and we want to be using .NET 9 SDK. Should I be forcing it to use .NET 9 with a global.json or some other setting?