@beeradmoore 2025-05-07
Ah, gotya. All good. 

I have my head in MAUI world a lot, [they follow latest release](https://dotnet.microsoft.com/en-us/platform/support/policy/maui) instead of LTS so I am not used to seeing .NET 8 pop up 😅