@beeradmoore 2025-05-07
Doesn't seem like numbers moved too much. Some higher, some lower (or within margin of error).

Keep in mind these tests are not using BenchmarkDotNet yet so it isn't doing the warmup and other things it does. Just loops over 10 times and I am jotting down the average.

## Debug (3.1.8)
| Device           | JpgLoad |  JpgResize | PngLoad | PngResize |
|------------------|---------|------------|---------|-----------|
| Android          | 1084.1  | 1312.4     | 37.1    | 44.4      |
| Android Emulator | 189.5   | 245.1      | 13.8    | 14.2      |

## Debug (Modernize JPEG Color Converters)
| Device           | JpgLoad |  JpgResize | PngLoad | PngResize |
|------------------|---------|------------|---------|-----------|
| Android          | 1344.77 | 1586.2     | 37.5    | 48.1      |
| Android Emulator | 233.3   | 285.8      | 13.8    | 15.6      |

## Debug (this updated PR)
| Device           | JpgLoad |  JpgResize | PngLoad | PngResize |
|------------------|---------|------------|---------|-----------|
| Android          | 1366.6  | 1605.1     | 37.1    | 48.4      |
| Android Emulator | 245.5   | 287.8      | 15.5    | 15.0      |

## Release (3.1.8)
| Device           | JpgLoad |  JpgResize | PngLoad | PngResize |
|------------------|---------|------------|---------|-----------|
| Android          | 285.5   | 392.9      | 19.3    | 26.1      |
| Android Emulator | 83.7    | 96.2       | 9.4     | 9.9       |

## Release (Modernize JPEG Color Converters)
| Device           | JpgLoad |  JpgResize | PngLoad | PngResize |
|------------------|---------|------------|---------|-----------|
| Android          | 341.0   | 469.4      | 20.1    | 25.8      |
| Android Emulator | 99.3    | 121.1      | 9.2     | 10.4      |

## Release (this updated PR)
| Device           | JpgLoad |  JpgResize | PngLoad | PngResize |
|------------------|---------|------------|---------|-----------|
| Android          | 360.3   | 482.0      | 19.5    | 27.5      |
| Android Emulator | 106.1   | 119.9      | 10.5    | 12.6      |