@JimBobSquarePants 2025-05-07
I think we need to find a way to properly benchmark because I cannot see how the numbers could be worse in this PR than the last one.