@beeradmoore 2025-05-07
New project I started is using that with MAUI. I'm having some issues getting it working with Mac Catalyst (macOS desktop) variant.

But if the current issues are Android I can just do a net9.0-android repo with Android only app to put the tests in to focus on that while I deal with full MAUI later