@JimBobSquarePants 2025-05-07
Yeah, as I recall iOS was very good, let’s focus on the numbers for Android