@tannergooding 2025-05-07
How is the project being compiled for Android, is it using Mono LLVM or standard Mono AOT?