@beeradmoore 2025-05-07
@tannergooding , the Android test app the numbers are coming from can have its guts found [here](https://github.com/beeradmoore/ImageSharpMAUITest/blob/main/ImageSharpMAUITest/ImageSharpMAUITest/ImageSharpMAUITest.csproj#L74-L83). AOT+LLVM (not using profiled AOT) for release mode.

For compiling/deploying/running it I used [this](https://github.com/beeradmoore/ImageSharpMAUITest/blob/main/ImageSharpMAUITest/run_android.sh) script.

So something like, 
```
dotnet build \
    /t:Run \
    --framework net9.0-android \
    --configuration Release\
    -p:RuntimeIdentifier= android-arm64\
    ImageSharpMAUITest/ImageSharpMAUITest.csproj
```

As for the nuget itself I put in the tests I am just using `dotnet pack -c Release`