@tannergooding 2025-05-07
Just for a test, could you try it without LLVM as well? That is, with `<EnableLLVM>false</EnableLLVM>`?