@tannergooding 2025-05-07
For reference, these are the accelerated APIs in Mono on .NET 8...

https://github.com/dotnet/runtime/blob/v8.0.0/src/mono/mono/mini/simd-intrinsics.c#L2381-L2399 for APIs defined on `Vector128<T>`

https://github.com/dotnet/runtime/blob/v8.0.0/src/mono/mono/mini/simd-intrinsics.c#L1149-L1219 for APIs defined on `Vector128` (non-generic static class)

You can see that for the first group (https://github.com/dotnet/runtime/blob/v8.0.0/src/mono/mono/mini/simd-intrinsics.c#L2403C1-L2588) it has paths for V128 to work with or without LLVM (it is entered from https://github.com/dotnet/runtime/blob/v8.0.0/src/mono/mono/mini/simd-intrinsics.c#L5821-L5824)

The second group has something similar: https://github.com/dotnet/runtime/blob/v8.0.0/src/mono/mono/mini/simd-intrinsics.c#L1369-L2379 -- It is entered from https://github.com/dotnet/runtime/blob/v8.0.0/src/mono/mono/mini/simd-intrinsics.c#L5816-L5819

Since it should work with or without LLVM, I'd at least like to determine if one of them appears to be working as intended. That can potentially help root cause the issue and what needs to be looked at next.