@beeradmoore 2025-05-07
## Release (this PR), Android device
| EnableLLVM | JpgLoad |  JpgResize | PngLoad | PngResize |
|------------|---------|------------|---------|-----------|
| true       | 292.3   | 428.9      | 21.0    | 26.0      |
| false      | 685.5   | 960.3      | 33.1    | 49.2      |